fix: Added condition to delete socketIo in socket-io.config.js.

+++ config/socket-io.config.js
@@ -14,6 +14,10 @@ const initializeSocketIo = (httpServer, CorsOrigin) => {
   io.on('connection', function(socket) {
     socket.on('setSocketId', async ({userId, socketId, oldSocketId}) => {
       try {
+        console.log('ICI----------------------------')
+        console.log(userId);
+        console.log(socketId);
+        console.log(oldSocketId);
         let socketIoDb = await SocketIoModel.findOne({userId : userId});
         if(socketIoDb){
           let arraySocketId = socketIoDb.socketId;
@@ -43,15 +47,17 @@ const initializeSocketIo = (httpServer, CorsOrigin) => {
     socket.on('disconnect', async () => {
       try {
         let socketIoDb = await SocketIoModel.findOne({socketId : socket.id});
-        let arraySocketId = socketIoDb.socketId;
-        if(arraySocketId.length > 1){
-          const indexSocketId = arraySocketId.findIndex(socketId => socketId === socket.id); 
-          arraySocketId.splice(indexSocketId, 1);
-          await SocketIoModel.findByIdAndUpdate(socketIoDb._id, { socketId: arraySocketId }, { override: true, upsert: true, new: true });
-        }else{
-          await SocketIoModel.findOneAndDelete({
-            socketId : socket.id,
-          });
+        if(socketIoDb){
+          let arraySocketId = socketIoDb.socketId;
+          if(arraySocketId.length > 1){
+            const indexSocketId = arraySocketId.findIndex(socketId => socketId === socket.id); 
+            arraySocketId.splice(indexSocketId, 1);
+            await SocketIoModel.findByIdAndUpdate(socketIoDb._id, { socketId: arraySocketId }, { override: true, upsert: true, new: true });
+          }else{
+            await SocketIoModel.findOneAndDelete({
+              socketId : socket.id,
+            });
+          }
         }
       } catch (error) {
         if(env.toUpperCase() === environments.PRODUCTION){
